(GH-813) Abstract Powershell shell implementation
This commit factors out the Powershell code from the WinRM and Windows
local transports to match the transport/shell split implemented for SSH
and Bash. That is, the transports are now merely responsible for
connecting to the target, executing commands and uploading files, while
the powershell shell class handles running tasks, etc.

We no longer have a dedicated local_windows transport, instead using the
standard local transport. The local transport will run any command it's
given via powershell. This matches the non-windows behavior where
commands are run via shell.

Due to bugs in passing stdin from ruby to powershell, we cannot simply
send the commands over stdin. Therefore, we write the commands to a file
and point powershell at that. This also avoids limitations around
command-line length if we were to instead run via `powershell.exe -Command`.

Because commands are now run via powershell, we can run into problems
with powershell inserting a byte order marker in content that we supply
to a task via stdin. For that reason, we now detect whether the
[Console] object is configured to use UTF-8 and, if so, we replace its
encoding a non-BOM version of UTF-8.

This also standardizes the newline behavior between the local and WinRM
transports. Previously, the WinRM transport would return \r\n while the
local transport would return \n. They now consistently return \r\n.

!feature

* **Commands run over local transport on Windows use powershell**
  ([#1708](https://github.com/puppetlabs/bolt/pull/1708))

  Previously, the local transport on Windows would exec commands
  directly, meaning powershell constructs couldn't be used. These
  commands are now always executed through powershell, so powershell
  commands and script snippets can be run.

* **Commands and tasks on Windows now consistently return \r\n**
  ([#1708](https://github.com/puppetlabs/bolt/pull/1708))

  The local transport on Windows was returning \n while WinRM returned
  \r\n. They are now consistent and always use \r\n.
