+++ spec/shared_examples/transport.rb
@@ -38,15 +38,15 @@ end
 
 def windows_context
   {
-    stdout_command: ['echo hello', /^hello$/],
+    stdout_command: ['echo hello', /^hello\r$/],
     stderr_command: ['echo oops 1>&2', /oops/],
     destination_dir: 'C:/mytmp',
     supported_req: 'powershell',
     extension: '.ps1',
     unsupported_req: 'shell',
     cat_cmd: 'cat',
-    rm_cmd: 'rm -rf',
-    ls_cmd: 'ls',
+    rm_cmd: 'rm -Recurse -Force',
+    ls_cmd: 'ls -Name',
     env_task: "Write-Output \"${env:PT_message_one}\n${env:PT_message_two}\"",
     stdin_task: "$line = [Console]::In.ReadLine()\nWrite-Output \"$line\"",
     find_task: 'Get-ChildItem -Path $env:PT__installdir -Recurse -File | % { Write-Host $_.Length $_.FullName  }',
@@ -134,7 +134,7 @@ shared_examples 'transport api' do
         expect(result.object).to eq(file.path)
 
         expect(
-          runner.run_command(target, "#{os_context[:cat_cmd]} #{remote_path}").value['stdout']
+          runner.run_command(target, "#{os_context[:cat_cmd]} #{remote_path}").value['stdout'].chomp
         ).to eq(contents)
 
         runner.run_command(target, "#{os_context[:rm_cmd]} #{remote_path}")
@@ -152,11 +152,12 @@ shared_examples 'transport api' do
         runner.upload(target, dir, target_dir)
 
         expect(
-          runner.run_command(target, "#{os_context[:ls_cmd]} #{target_dir}")['stdout'].split("\n")
-        ).to eq(%w[content subdir])
+          runner.run_command(target, "#{os_context[:ls_cmd]} #{target_dir}")['stdout'].lines.map(&:chomp)
+        ).to contain_exactly('content', 'subdir')
 
         expect(
-          runner.run_command(target, "#{os_context[:ls_cmd]} #{File.join(target_dir, 'subdir')}")['stdout'].split("\n")
+          runner.run_command(target, "#{os_context[:ls_cmd]} #{File.join(target_dir, 'subdir')}")['stdout']
+            .lines.map(&:chomp)
         ).to eq(%w[more])
 
         runner.run_command(target, "#{os_context[:rm_cmd]} #{target_dir}")
@@ -166,7 +167,7 @@ shared_examples 'transport api' do
 
   context 'run_script' do
     it "can run a script remotely" do
-      with_tempfile_containing('script test', os_context[:echo_script]) do |file|
+      with_tempfile_containing('script test', os_context[:echo_script], os_context[:extension]) do |file|
         result = runner.run_script(target, file.path, [])
         expect(result['stdout'].strip).to eq('')
         expect(result.action).to eq('script')
@@ -176,6 +177,15 @@ shared_examples 'transport api' do
 
     it "can run a script remotely with quoted arguments" do
       with_tempfile_containing('script-test-docker-quotes', os_context[:echo_script], os_context[:extension]) do |file|
+        expected = <<~QUOTED
+                     nospaces
+                     with spaces
+                     "double double"
+                     'double single'
+                     double "double" double
+                     double 'single' double
+                   QUOTED
+
         expect(
           runner.run_script(target,
                             file.path,
@@ -185,14 +195,8 @@ shared_examples 'transport api' do
                              "'double single'",
                              "double \"double\" double",
                              "double 'single' double"])['stdout']
-        ).to eq(<<~QUOTED)
-                  nospaces
-                  with spaces
-                  "double double"
-                  'double single'
-                  double "double" double
-                  double 'single' double
-                QUOTED
+            .lines.map(&:chomp)
+        ).to eq(expected.lines.map(&:chomp))
       end
     end
 
@@ -201,8 +205,8 @@ shared_examples 'transport api' do
                    make_sensitive('$ecret!')]
       with_tempfile_containing('sensitive_test', os_context[:echo_script], os_context[:extension]) do |file|
         expect(
-          runner.run_script(target, file.path, arguments)['stdout']
-        ).to eq("non-sensitive-arg\n$ecret!\n")
+          runner.run_script(target, file.path, arguments)['stdout'].lines.map(&:chomp)
+        ).to eq(%w[non-sensitive-arg $ecret!])
       end
     end
 
@@ -211,10 +215,8 @@ shared_examples 'transport api' do
         expect(
           runner.run_script(target,
                             file.path,
-                            ['echo $HOME; cat /etc/passwd'])['stdout']
-        ).to eq(<<~SHELLWORDS)
-        echo $HOME; cat /etc/passwd
-        SHELLWORDS
+                            ['echo $HOME; cat /etc/passwd'])['stdout'].chomp
+        ).to eq("echo $HOME; cat /etc/passwd")
       end
     end
   end
@@ -224,7 +226,7 @@ shared_examples 'transport api' do
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_task_containing('tasks_test', os_context[:env_task], 'environment', os_context[:extension]) do |task|
         result = runner.run_task(target, task, arguments)
-        expect(result.message).to eq("Hello from task\nGoodbye\n")
+        expect(result.message.chomp).to eq("Hello from task\nGoodbye")
         expect(result.object).to eq('tasks_test')
       end
     end
@@ -249,7 +251,7 @@ shared_examples 'transport api' do
       content = "#{os_context[:env_task]}\n#{os_context[:stdin_task]}"
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_task_containing('tasks-test-both', content, 'both', os_context[:extension]) do |task|
-        expect(runner.run_task(target, task, arguments).message.strip).to eq(<<~OUTPUT.strip)
+        expect(runner.run_task(target, task, arguments).message.lines.map(&:chomp)).to eq(<<~OUTPUT.lines.map(&:chomp))
         Hello from task
         Goodbye
         {"message_one":"Hello from task","message_two":"Goodbye"}
@@ -276,7 +278,7 @@ shared_examples 'transport api' do
       arguments = { 'message_one' => make_sensitive('$ecret!'),
                     'message_two' => make_sensitive(deep_hash) }
       with_task_containing('tasks_test_sensitive', os_context[:env_task], 'both', os_context[:extension]) do |task|
-        expect(runner.run_task(target, task, arguments).message).to eq(<<~SHELL)
+        expect(runner.run_task(target, task, arguments).message.lines.map(&:chomp)).to eq(<<~SHELL.lines.map(&:chomp))
         $ecret!
         {"k":"v","arr":[1,2,3]}
         SHELL
@@ -300,8 +302,8 @@ shared_examples 'transport api' do
       with_task_containing('tasks_test', contents, 'environment', os_context[:extension]) do |task|
         reqs = [os_context[:supported_req]]
         task.metadata['implementations'] = [{ 'name' => 'tasks_test', 'requirements' => reqs }]
-        expect(runner.run_task(target, task, arguments).message)
-          .to eq("Hello from task\nGoodbye\n")
+        expect(runner.run_task(target, task, arguments).message.chomp)
+          .to eq("Hello from task\nGoodbye")
       end
     end
 
@@ -467,9 +469,10 @@ shared_examples 'transport api' do
     let(:tmpdir) { File.join(os_context[:destination_dir], 'mytempdir') }
 
     it "errors when tmpdir doesn't exist" do
+      skip "Windows will create the directory anyway" if Bolt::Util.windows?
       set_config(target, 'tmpdir' => tmpdir)
 
-      with_tempfile_containing('script dir', 'dummy script') do |file|
+      with_tempfile_containing('script dir', 'dummy script', os_context[:extension]) do |file|
         expect {
           runner.run_script(target, file.path, [])
         }.to raise_error(Bolt::Node::FileError, /Could not make tempdir.*#{Regexp.escape(tmpdir)}/)
